refactor(sequence): move chans to build/extract
Refactor the parts of the chan-centered package into the build and
extract packages as needed.

Also to match naming, change all the As* functions into To* equivalents.
